Move core of WordPress Service to Trait
